requirements.txt
"unstructured[all-docs]"
pillow
pydantic
lxml
matplotlib
langchain
openai
python-dotenv
fastapi
jinja2
python-multipart
uvicorn
tiktoken
pydantic 
lxml 
openai 
chromadb 
tiktoken 
opencv-python
langchain_core
langchain_openai
langchain
chromadb
langchain-google-genai
langchain_community